+++ rvm/src/org/jikesrvm/compilers/baseline/SwitchBranchProfile.java
@@ -33,7 +33,7 @@ public final class SwitchBranchProfile extends BranchProfile {
     super(_bci, sumCounts(cs, start, numEntries));
     counts = new float[numEntries];
     for (int i = 0; i < numEntries; i++) {
-      counts[i] = (float) cs[start + i];
+      counts[i] = cs[start + i];
     }
   }
 
@@ -65,7 +65,7 @@ public final class SwitchBranchProfile extends BranchProfile {
   private static float sumCounts(int[] counts, int start, int numEntries) {
     float sum = 0.0f;
     for (int i = start; i < start + numEntries; i++) {
-      sum += (float) counts[i];
+      sum += counts[i];
     }
     return sum;
   }
