+++ rvm/src/org/jikesrvm/osr/VariableElement.java
@@ -104,7 +104,7 @@ public class VariableElement implements OSRConstants {
     this.kind = what_kind;
     this.num = (char)which_num;
     this.tcode = type;
-    this.value = (long) ibits & 0x0FFFFFFFFL;
+    this.value = ibits & 0x0FFFFFFFFL;
     this.ref = null;
   }
 
@@ -147,7 +147,7 @@ public class VariableElement implements OSRConstants {
     this.num = (char)which_num;
     this.tcode = type;
     if (VM.BuildFor32Addr) {
-      this.value = ((long) word.toInt()) & 0x0FFFFFFFFL;
+      this.value = (word.toInt()) & 0x0FFFFFFFFL;
     } else {
       this.value = word.toLong();
     }
