+++ rvm/src/org/jikesrvm/osr/ia32/OptExecutionStateExtractor.java
@@ -400,7 +400,7 @@ public abstract class OptExecutionStateExtractor extends ExecutionStateExtractor
     // for LCONST type, the value is the value
     if (vtypeLow == LCONST || vtypeLow == ACONST) {
       if (VM.VerifyAssertions) VM._assert(vtypeHigh == vtypeLow);
-      return ((((long) valueHigh) << 32) | (((long) valueLow) & 0x0FFFFFFFFL));
+      return ((((long) valueHigh) << 32) | ((valueLow) & 0x0FFFFFFFFL));
 
     } else if (VM.BuildFor32Addr) {
       if (VM.VerifyAssertions) VM._assert(vtypeHigh == PHYREG || vtypeHigh == SPILL);
@@ -410,15 +410,15 @@ public abstract class OptExecutionStateExtractor extends ExecutionStateExtractor
       long lowPart, highPart;
 
       if (vtypeLow == PHYREG) {
-        lowPart = ((long)registers.gprs.get(valueLow).toInt()) & 0x0FFFFFFFFL;
+        lowPart = (registers.gprs.get(valueLow).toInt()) & 0x0FFFFFFFFL;
       } else {
-        lowPart = ((long)Magic.getIntAtOffset(stack, fpOffset.minus(valueLow))) & 0x0FFFFFFFFL;
+        lowPart = (Magic.getIntAtOffset(stack, fpOffset.minus(valueLow))) & 0x0FFFFFFFFL;
       }
 
       if (vtypeHigh == PHYREG) {
-        highPart = ((long)registers.gprs.get(valueHigh).toInt());
+        highPart = (registers.gprs.get(valueHigh).toInt());
       } else {
-        highPart = ((long)Magic.getIntAtOffset(stack, fpOffset.minus(valueHigh)));
+        highPart = (Magic.getIntAtOffset(stack, fpOffset.minus(valueHigh)));
       }
 
       return (highPart << 32) | lowPart;
