+++ rvm/src/org/jikesrvm/runtime/Time.java
@@ -27,7 +27,7 @@ public class Time {
    * @return c converted to milliseconds
    */
   public static double nanosToMillis(long c) {
-    return ((double)c)/1e6;
+    return (c)/1e6;
   }
 
   /**
@@ -62,7 +62,7 @@ public class Time {
   }
 
   public static double nanosToSecs(long nanos) {
-    return ((double)nanos)/1E9;
+    return (nanos)/1E9;
   }
 
   public static long secsToNanos(double secs) {
